pyproject.toml
[tool.poetry]
name = "afrourban"
version = "0.1.0"
description = "Template for Django projects"
authors = ["Ernest Jumbe <e2jeyy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "5.1.2"
psycopg2-binary = "2.9.9"
djangorestframework = "^3.17.1"
djangorestframework-simplejwt = "^5.5.1"
pillow = "^12.1.1"
python-dotenv = "^1.2.2"
structlog = "^25.2.0"

[tool.poetry.group.dev.dependencies]
django-extensions = "3.2.3"
django-debug-toolbar = "4.4.6"
werkzeug = "^3.1.3"
mypy = "^1.15.0"
factory-boy = "^3.3.3"
pytest-django = "^4.12.0"


[tool.poetry.group.prd.dependencies]
gunicorn = "^23.0.0"
django-robots = "^6.1"
django-htmlmin = "^0.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disable_error_code = ["import-untyped"]
